+++ griptape/config/openai_structure_config.py
@@ -11,11 +11,11 @@ from griptape.drivers import (
     OpenAiChatPromptDriver,
     OpenAiEmbeddingDriver,
     OpenAiImageGenerationDriver,
-    OpenAiVisionImageQueryDriver,
     BaseTextToSpeechDriver,
     OpenAiTextToSpeechDriver,
     BaseAudioTranscriptionDriver,
     OpenAiAudioTranscriptionDriver,
+    OpenAiImageQueryDriver,
 )
 
 
@@ -30,9 +30,7 @@ class OpenAiStructureConfig(StructureConfig):
         metadata={"serializable": True},
     )
     image_query_driver: BaseImageQueryDriver = field(
-        default=Factory(lambda: OpenAiVisionImageQueryDriver(model="gpt-4-vision-preview")),
-        kw_only=True,
-        metadata={"serializable": True},
+        default=Factory(lambda: OpenAiImageQueryDriver(model="gpt-4o")), kw_only=True, metadata={"serializable": True}
     )
     embedding_driver: BaseEmbeddingDriver = field(
         default=Factory(lambda: OpenAiEmbeddingDriver(model="text-embedding-3-small")),