+++ griptape/drivers/image_generation/dummy_image_generation_driver.py
@@ -7,7 +7,7 @@ from griptape.exceptions import DummyException
 
 @define
 class DummyImageGenerationDriver(BaseImageGenerationDriver):
-    model: str = field(init=False)
+    model: None = field(init=False, default=None, kw_only=True)
 
     def try_text_to_image(self, prompts: list[str], negative_prompts: Optional[list[str]] = None) -> ImageArtifact:
         raise DummyException(__class__.__name__, "try_text_to_image")